+++ src/constellation/__init__.py
@@ -2,4 +2,4 @@
 
 __all__ = ["__version__"]
 
-__version__ = "5.4.0"
+__version__ = "5.5.0"